Move auth functions to separate helper
