+++ cmd/auth.go
@@ -16,78 +16,10 @@ limitations under the License.
 package cmd
 
 import (
-	"fmt"
-	"net/http"
-	"os"
-
 	"github.com/dvdmuckle/goify/cmd/helper"
-	"github.com/golang/glog"
-	spotifyAuth "github.com/markbates/goth/providers/spotify"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"github.com/zmb3/spotify"
 )
-
-const redirectURI = "http://localhost:8888/callback"
-
-var (
-	authenticator = spotify.NewAuthenticator(redirectURI)
-	ch            = make(chan *spotify.Client)
-	clientID      string
-	secret        string
-	state         = "ringdingthing"
-)
-
-func auth() {
-	clientID = viper.GetString("spotifyclientid")
-	secret = viper.GetString("spotifysecret")
-	if clientID == "" || secret == "" {
-		fmt.Println("Please configure your Spotify client ID and secret in the config file at ~/.config/goify/config.yaml")
-		os.Exit(1)
-	}
-	//TODO: Go back and reimplement this with the spotifyAuth library
-	provider := spotifyAuth.New(clientID, secret, redirectURI)
-	if viper.GetString("auth.accesstoken") != "" && provider.RefreshTokenAvailable() {
-		viper.Set("auth.accesstoken", helper.RefreshToken(provider, viper.GetString("auth.refreshtoken")))
-	} else {
-		fmt.Println("Getting token...")
-		authenticator.SetAuthInfo(clientID, secret)
-		http.HandleFunc("/callback", completeAuth)
-		go http.ListenAndServe(":8888", nil)
-		url := authenticator.AuthURL(state)
-		fmt.Println("Please log in to Spotify by clicking the following link:", url)
-		//wait for auth to finish
-		client := <-ch
-
-		user, err := client.CurrentUser()
-		if err != nil {
-			glog.Fatal(err)
-		}
-		token, err := client.Token()
-		if err != nil {
-			glog.Fatal(err)
-		}
-		viper.Set("auth", token)
-		viper.WriteConfigAs(cfgFile)
-		fmt.Println("Login successful as", user.ID)
-	}
-}
-
-func completeAuth(w http.ResponseWriter, r *http.Request) {
-	tok, err := authenticator.Token(state, r)
-	if err != nil {
-		http.Error(w, "Couldn't get token", http.StatusForbidden)
-		glog.Fatal(err)
-	}
-	if st := r.FormValue("state"); st != state {
-		http.NotFound(w, r)
-		glog.Fatalf("State mismatch: %s != %s\n", st, state)
-	}
-	// use the token to get an authenticated client
-	client := authenticator.NewClient(tok)
-	fmt.Fprintf(w, "Login Completed!")
-	ch <- &client
-}
 
 // authCmd represents the auth command
 var authCmd = &cobra.Command{
@@ -96,7 +28,7 @@ var authCmd = &cobra.Command{
 	Long: `Authenticates with Spotify by printout out a login link, which will then save your access token to the config file
 	       Use this command after the initial login to refresh your access token`,
 	Run: func(cmd *cobra.Command, args []string) {
-		auth()
+		helper.Auth(viper.GetViper(), cfgFile)
 	},
 }
